Refactor contributorName shortcode as fullName filter
